notification feature added

+++ Server/controller/notificationController.js
@@ -3,10 +3,9 @@ const Notification = require("../models/notifications");
 
 const getAllNotifications = async (req, res) => {
     try {
-        const {usertype} = req.query;
-        console.log("usetyppe",usertype);
-
-        const notifications = await Notification.find({usertype}).sort({ createdAt: -1 });
+        const {usertype,userId} = req.query;
+        console.log("usetyppe",usertype,userId);
+        const notifications = await Notification.find({usertype,userId}).sort({ createdAt: -1 });
         res.status(200).json({notifications});
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -15,10 +14,10 @@ const getAllNotifications = async (req, res) => {
 
 const getTotalUnreadNotifications = async (req, res) => {
     try {
-        const {usertype} = req.query;
-        console.log("usetyppe",usertype);
-        const notifications = await Notification.find({usertype}).sort({ createdAt: -1 });
-        const totalUnread = await Notification.countDocuments({ read: false , usertype });
+        const {usertype,userId} = req.query;
+        console.log("usetyppe",usertype,"userId",userId);
+        const notifications = await Notification.find({usertype,userId}).sort({ createdAt: -1 });
+        const totalUnread = await Notification.countDocuments({ read: false , usertype,userId });
         res.status(200).json({ totalUnread,notifications });
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -27,13 +26,14 @@ const getTotalUnreadNotifications = async (req, res) => {
 
 const markAsRead = async (req, res) => {
     const { id } = req.params;
-
+    const {usertype,userId} = req.query;
     try {
         const notification = await Notification.findByIdAndUpdate(id, { read: true }, { new: true });
         if (!notification) {
             return res.status(404).json({ error: 'Notification not found' });
         }
-        res.status(200).json(notification);
+        const notifications = await Notification.find({usertype,userId}).sort({ createdAt: -1 });
+        res.status(200).json({notifications});
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -41,14 +41,14 @@ const markAsRead = async (req, res) => {
 
 const deleteNotification = async (req, res) => {
     const { id } = req.params;
-    const {usertype} = req.query;
+    const {usertype,userId} = req.query;
 
     try {
         const notification = await Notification.findByIdAndDelete(id);
         if (!notification) {
             return res.status(404).json({ message: 'Notification not found' });
         }
-        const notifications = await Notification.find({usertype}).sort({ createdAt: -1 });
+        const notifications = await Notification.find({usertype,userId}).sort({ createdAt: -1 });
         res.status(200).json({ message: 'Notification deleted successfully',notifications });
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -57,8 +57,8 @@ const deleteNotification = async (req, res) => {
 
 const deleteAllNotifications = async (req, res) => {
     try {
-        const {usertype} = req.query;
-        await Notification.deleteMany({usertype});
+        const {usertype,userId} = req.query;
+        await Notification.deleteMany({usertype,userId});
         res.status(200).json({ message: 'All notifications deleted successfully' });
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -68,11 +68,11 @@ const deleteAllNotifications = async (req, res) => {
 const readAllNotifications = async (req, res) => {
     try {
         // Extract the usertype from the query parameters
-        const { usertype } = req.query;
+        const { usertype,userId } = req.query;
 
         // Update all notifications for the given usertype to mark them as read
-        await Notification.updateMany({ usertype }, { $set: { read: true } });
-        const notifications = await Notification.find({usertype}).sort({ createdAt: -1 });
+        await Notification.updateMany({ usertype,userId }, { $set: { read: true } });
+        const notifications = await Notification.find({usertype,userId}).sort({ createdAt: -1 });
